accuracy function to metrics file
